Minor bugfixes and tweaks.

+++ CameraControls.py
@@ -260,7 +260,7 @@ class CameraThread(threading.Thread):
                             if not video_created:
                                 video, experiment_info_name, exp_info_params = self.create_video_writer()
                                 video_created = True
-                            video.write(c_p['image'])
+                            video.write(np.uint8(c_p['image']))
                         # Capture an image and update the image count
                         image_count = image_count+1
